requirements.txt
langchain-google-genai
fastapi
uvicorn
supabase
python-dotenv
pydantic
langchain
langchain-core
langgraph
httpx
python-multipart
